+++ src/translations/en/neo.php
@@ -114,4 +114,12 @@ return [
     'Edit {blockType} blocks' => 'Edit {blockType} blocks',
     'Create blocks' => 'Create blocks',
     'Delete blocks' => 'Delete blocks',
+    'Collapse all blocks?' => 'Collapse all blocks?',
+    'Whether all blocks should display as collapsed when loading an element editor.' => 'Whether all blocks should display as collapsed when loading an element editor.',
+    'Default ’Always Show Dropdown’ setting' => 'Default ’Always Show Dropdown’ setting',
+    'Whether block type groups will always show as a dropdown menu, if their ’Always Show Dropdown’ setting is set to use the global setting.' => 'Whether block type groups will always show as a dropdown menu, if their ’Always Show Dropdown’ setting is set to use the global setting.',
+    'Optimise search indexing?' => 'Optimise search indexing?',
+    'Whether to skip updating search indexes for Neo blocks that have no sub-fields set to use their values as search keywords, or that belong to Neo fields that aren’t set to use the field’s values as search keywords.' => 'Whether to skip updating search indexes for Neo blocks that have no sub-fields set to use their values as search keywords, or that belong to Neo fields that aren’t set to use the field’s values as search keywords.',
+    'Use the new block grid?' => 'Use the new block grid?',
+    'Whether to use the new block grid, instead of new block buttons and dropdowns.' => 'Whether to use the new block grid, instead of new block buttons and dropdowns.',
 ];